chore: move constants into config folder
