+++ src/drawings/example/Example.jsx
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { CanvasBuilder } from '@/components/Canvas'
-import { Color } from '@/constants'
+import { Color } from '@/config'
 
 /** @type {import('@/components/Canvas/CanvasBuilder').DrawFactory} */
 function drawExample (canvas, { scale }) {
